pyproject.toml
[tool.poetry]
name = "finalproject"
version = "0.1.0"
description = ""
authors = ["Самир Багиров <samir_261999@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.14.0"
django-filter = "^24.1"
drf-spectacular = "^0.27.1"
pillow = "^10.2.0"
pyyaml = "^6.0.1"
gunicorn = "^21.2.0"
frontend = "^0.0.3"
django-js-asset = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
